api/requirements.txt

Flask==3.0.3
Flask-Cors==4.0.1
stytch==10.1.0
python-dotenv==1.0.1